src/c3ds/requirements/dev.txt
-r production.txt
pyproject-flake8
isort
django-debug-toolbar
django-extensions
pur
